Extract builds API sample into a separate subcommand
